feat: protocol CoreDataProcessing 생성하여 기존 메서드 분리
- 기존 DiaryListTableViewController에 있던 메서드 createDiaryContents와 fetchDiaryContents를
   create(text:)와 getContents<T>(_ request:) -> [T]?로 분리
